switch16.js
New protocol for wireless remote control switch "Intertek 22541673" which can be bought in Germany for example in "Bauhaus". It's a set of 3 switches with one remote. The switches are self learning. I have only 1 set so that I do not exactly know where the id begins. At least this protocol is working for my switches.