Weather14 dont emit an humidity value
In some cases the protocol provides an humidity value of 204. This was originally forced to 0 by @mwittig. But this 0 would stored in the log history of the device. To prevent this we add humidity only when an right value is received. I have only on problem, i cant add a field to an object without the extend function. Eventually there exist an better way or we move the function into the helper.